pyproject.toml
[build-system]
requires = ["setuptools>=61.2", "setuptools_scm[toml]>=3.4.1"]
build-backend = "setuptools.build_meta"

[project]
name = "cmdix"
authors = [
	{ name = "Jason R. Coombs", email = "jaraco@jaraco.com" },
]
description = "Unix commands in Pure Python"
readme = "README.rst"
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: Developers",
	"License :: OSI Approved :: MIT License",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3 :: Only",
	"Environment :: Console",
	"Topic :: System :: Shells",
	"Topic :: Utilities",
]
requires-python = ">=3.9"
dependencies = [
	"importlib_resources",
	'backports.hook_compressed; python_version < "3.10"',
	"more_itertools",
	"python-dateutil",
]
dynamic = ["version"]

[project.urls]
Source = "https://github.com/jaraco/cmdix"

[project.optional-dependencies]
test = [
	# upstream
	"pytest >= 6, != 8.1.*",

	# local
]

doc = [
	# upstream
	"sphinx >= 3.5",
	"jaraco.packaging >= 9.3",
	"rst.linker >= 1.9",
	"furo",
	"sphinx-lint",

	# local
]

check = [
	"pytest-checkdocs >= 2.4",
	"pytest-ruff >= 0.2.1; sys_platform != 'cygwin'",
]

cover = [
	"pytest-cov",
]

enabler = [
	"pytest-enabler >= 2.2",
]

type = [
	# upstream
	"pytest-mypy",

	# local
	"types-backports",
]


[project.scripts]
gunzip = "cmdix:run"
gzip = "cmdix:run"
uname = "cmdix:run"
tail = "cmdix:run"
wc = "cmdix:run"
tee = "cmdix:run"
base64 = "cmdix:run"
sha384sum = "cmdix:run"
env = "cmdix:run"
id = "cmdix:run"
uuidgen = "cmdix:run"
clear = "cmdix:run"
chmod = "cmdix:run"
mktemp = "cmdix:run"
sha1sum = "cmdix:run"
shuf = "cmdix:run"
cal = "cmdix:run"
tar = "cmdix:run"
sha224sum = "cmdix:run"
cat = "cmdix:run"
dirname = "cmdix:run"
sort = "cmdix:run"
httpd = "cmdix:run"
sh = "cmdix:run"
bzip2 = "cmdix:run"
mv = "cmdix:run"
ln = "cmdix:run"
chown = "cmdix:run"
sha512sum = "cmdix:run"
arch = "cmdix:run"
yes = "cmdix:run"
rmdir = "cmdix:run"
chroot = "cmdix:run"
ls = "cmdix:run"
cp = "cmdix:run"
logger = "cmdix:run"
rm = "cmdix:run"
pwd = "cmdix:run"
mkdir = "cmdix:run"
zip = "cmdix:run"
more = "cmdix:run"
whoami = "cmdix:run"
nl = "cmdix:run"
sha256sum = "cmdix:run"
kill = "cmdix:run"
crond = "cmdix:run"
sendmail = "cmdix:run"
basename = "cmdix:run"
touch = "cmdix:run"
bunzip2 = "cmdix:run"
diff = "cmdix:run"
sleep = "cmdix:run"
mount = "cmdix:run"
watch = "cmdix:run"
shred = "cmdix:run"
md5sum = "cmdix:run"
init = "cmdix:run"
uptime = "cmdix:run"
expand = "cmdix:run"
seq = "cmdix:run"
wget = "cmdix:run"


[tool.setuptools_scm]


[tool.pytest-enabler.mypy]
# Disabled due to jaraco/skeleton#143
